requirements.txt
pdfplumber
azure-functions